requirements.txt
pytorch
flask
oauthlib
numpy
arrow
